requirements.txt
asgiref==3.2.7
astroid==2.3.3
autopep8==1.5
beautifulsoup4==4.8.2
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
Django==3.0.5
django-cors-headers==3.2.1
djangorestframework==3.11.0
idna==2.9
isort==4.3.21
joblib==0.14.1
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.18.2
pandas==1.0.3
pycodestyle==2.5.0
pylint==2.4.4
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
scikit-surprise==1.1.0
scipy==1.4.1
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
typed-ast==1.4.1
urllib3==1.25.8
wrapt==1.11.2
